fix: reduce the file size below 400
