+++ lib/components/RunFrameWithApi/EnhancedFileSelectorCombobox/index.ts
@@ -1,0 +1,2 @@
+export { EnhancedFileSelectorCombobox } from "./EnhancedFileSelectorCombobox"
+export type { FileNode } from "./parseFilesToTree"
